requirements.txt
Flask==2.1.1
pytest==7.1.1
python-dotenv==0.20.0
PyYAML==6.0
